\name{grepV}
\alias{grepV}
\title{
get submatrix via row-names
}
\description{
simple pattern matching using \code{\link{grep}} 
}
\usage{
grepV(V, x,
      normalize = FALSE,
      simplifynames = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{V}{
target matrix (with rownames)
}
  \item{x}{
character string
}
\item{normalize}{
If \code{TRUE}, normalize such that the sum of the row-vectors is 1
}
\item{simplifynames}{
If \code{TRUE}, remove the pattern specified by \code{x} from rownames
}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}
