\name{meanV_array}
\alias{meanV_array}
\title{Compute Mean of Posterior Distributions}
\description{
  Computes the mean of Gamma-distributed variational parameters.
}
\usage{
meanV_array(out)
}
\arguments{
  \item{out}{List containing \code{shape} and \code{rate} matrices.}
}
\value{
  A list of matrices representing the means of the posterior Gamma distributions.
}
