\name{moltenNMF-package}
\alias{moltenNMF-package}
\alias{moltenNMF}
\docType{package}
\title{
\packageTitle{moltenNMF}
}
\description{
\packageDescription{moltenNMF}
}
\seealso{
Ko Abe, Teppei Shimamura, UNMF: a unified nonnegative matrix factorization for multi-dimensional omics data, Briefings in Bioinformatics, Volume 24, Issue 5, September 2023, bbad253, \url{https://doi.org/10.1093/bib/bbad253}
}
\author{
\packageAuthor{moltenNMF}

Maintainer: \packageMaintainer{moltenNMF}
}
\examples{
### see vignette
}
