\name{moltenPPCA-package}
\alias{moltenPPCA-package}
\alias{moltenPPCA}
\docType{package}
\title{
\packageTitle{moltenPPCA}
}
\description{
\packageDescription{moltenPPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moltenPPCA}
\packageIndices{moltenPPCA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{moltenPPCA}

Maintainer: \packageMaintainer{moltenPPCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
