\name{moltenNMF-package}
\alias{moltenNMF-package}
\alias{moltenNMF}
\docType{package}
\title{
\packageTitle{moltenNMF}
}
\description{
\packageDescription{moltenNMF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moltenNMF}
\packageIndices{moltenNMF}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{moltenNMF}

Maintainer: \packageMaintainer{moltenNMF}
}
\examples{
### see vignette
}
