\name{moltenPPCA-package}
\alias{moltenPPCA-package}
\alias{moltenPPCA}
\docType{package}
\title{
\packageTitle{moltenPPCA}
}
\description{
\packageDescription{moltenPPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moltenPPCA}
\packageIndices{moltenPPCA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{moltenPPCA}

Maintainer: \packageMaintainer{moltenPPCA}
}
\examples{
### see vignette
}
